requirements.txt
asgiref==3.11.0
Django==4.2.27
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
et_xmlfile==2.0.0
joblib==1.5.3
lxml==6.0.2
numpy==2.0.2
openpyxl==3.1.5
packaging==26.0
pandas==2.3.3
pillow==11.3.0
psycopg2-binary==2.9.11
PyJWT==2.10.1
PyPDF2==3.0.1
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-decouple==3.8
python-docx==1.2.0
pytz==2025.2
scikit-learn==1.6.1
scipy==1.13.1
six==1.17.0
sqlparse==0.5.5
threadpoolctl==3.6.0
typing_extensions==4.15.0
tzdata==2025.3
